+++ packages/commonjs/test/fixtures/function/indent-multiline-strings/main.js
@@ -1,0 +1,8 @@
+const dep = require('./dep.js');
+
+module.exports = `
+main
+multi
+line
+string
+`;