+++ packages/typescript/src/options/tsconfig.ts
@@ -23,6 +23,7 @@ import {
   DEFAULT_COMPILER_OPTIONS,
   EnumCompilerOptions,
   FORCED_COMPILER_OPTIONS,
+  OVERRIDABLE_EMIT_COMPILER_OPTIONS,
   PartialCompilerOptions
 } from './interfaces';
 import { normalizeCompilerOptions, makePathsAbsolute } from './normalize';
@@ -38,6 +39,10 @@ export interface TypeScriptConfig {
   errors: Diagnostic[];
   wildcardDirectories?: MapLike<WatchDirectoryFlags> | undefined;
   compileOnSave?: boolean | undefined;
+}
+
+function makeForcedCompilerOptions(noForceEmit: boolean) {
+  return { ...FORCED_COMPILER_OPTIONS, ...(noForceEmit ? {} : OVERRIDABLE_EMIT_COMPILER_OPTIONS) };
 }
 
 /**
@@ -134,7 +139,8 @@ const configCache = new Map() as import('typescript').Map<ExtendedConfigCacheEnt
 export function parseTypescriptConfig(
   ts: typeof import('typescript'),
   tsconfig: RollupTypescriptOptions['tsconfig'],
-  compilerOptions: PartialCompilerOptions
+  compilerOptions: PartialCompilerOptions,
+  noForceEmit: boolean
 ): TypeScriptConfig {
   /* eslint-disable no-undefined */
   const cwd = process.cwd();
@@ -161,7 +167,7 @@ export function parseTypescriptConfig(
         },
         ts.sys,
         basePath,
-        { ...compilerOptions, ...FORCED_COMPILER_OPTIONS },
+        { ...compilerOptions, ...makeForcedCompilerOptions(noForceEmit) },
         tsConfigPath,
         undefined,
         undefined,
@@ -181,7 +187,7 @@ export function parseTypescriptConfig(
         },
         ts.sys,
         basePath,
-        FORCED_COMPILER_OPTIONS,
+        makeForcedCompilerOptions(noForceEmit),
         tsConfigPath,
         undefined,
         undefined,