+++ packages/typescript/test/test.js
@@ -134,6 +134,30 @@ test.serial('ensures multiple outputs can be built', async (t) => {
   ]);
 });
 
+test.serial('supports emitting types also for single file output', async (t) => {
+  // Navigate to folder and use default local tsconfig instead of specifying tsconfig via file path
+  // as that would have the side effect that the tsconfig's path would be used as fallback path for
+  // the here unspecified outputOptions.dir, in which case the original issue wouldn't show.
+  process.chdir('fixtures/basic');
+
+  const warnings = [];
+  const bundle = await rollup({
+    input: 'main.ts',
+    plugins: [typescript({ declaration: true, declarationDir: 'dist' })],
+    onwarn(warning) {
+      warnings.push(warning);
+    }
+  });
+  // generate a single output bundle, in which case, declaration files were not correctly emitted
+  const output = await getCode(bundle, { format: 'esm', file: 'dist/main.js' }, true);
+
+  t.deepEqual(
+    output.map((out) => out.fileName),
+    ['main.js', 'main.d.ts']
+  );
+  t.is(warnings.length, 0);
+});
+
 test.serial('relative paths in tsconfig.json are resolved relative to the file', async (t) => {
   const bundle = await rollup({
     input: 'fixtures/relative-dir/main.ts',
@@ -1219,4 +1243,67 @@ test.serial('correctly resolves types in a nodenext module', async (t) => {
   t.true(code.includes('const bar = foo'), code);
   t.is(warnings.length, 1);
   t.is(warnings[0].code, 'UNRESOLVED_IMPORT');
+});
+
+test.serial('noForceEmit option defers to tsconfig.json for emitDeclarationOnly', async (t) => {
+  const input = 'fixtures/noForceEmit/emitDeclarationOnly/main.ts';
+  const warnings = [];
+  const bundle = await rollup({
+    input,
+    plugins: [
+      typescript({
+        tsconfig: 'fixtures/noForceEmit/emitDeclarationOnly/tsconfig.json',
+        noForceEmit: true
+      })
+    ],
+    onwarn(warning) {
+      warnings.push(warning);
+    }
+  });
+  // generate a single output bundle, in which case, declaration files were not correctly emitted
+  const output = await getCode(
+    bundle,
+    { format: 'esm', file: 'fixtures/noForceEmit/emitDeclarationOnly/dist/main.js' },
+    true
+  );
+
+  t.deepEqual(
+    output.map((out) => out.fileName),
+    // original file is passed through, main.d.ts is emitted
+    ['main.js', 'main.d.ts']
+  );
+  t.is(warnings.length, 0);
+  // test that NO transpilation happened
+  const originalCode = fs.readFileSync(path.join(__dirname, input), 'utf8');
+  t.is(output[0].code, originalCode);
+});
+
+test.serial('noForceEmit option defers to tsconfig.json for noEmit', async (t) => {
+  const input = 'fixtures/noForceEmit/noEmit/main.ts';
+  const warnings = [];
+  const bundle = await rollup({
+    input,
+    plugins: [
+      typescript({ tsconfig: 'fixtures/noForceEmit/noEmit/tsconfig.json', noForceEmit: true })
+    ],
+    onwarn(warning) {
+      warnings.push(warning);
+    }
+  });
+  // generate a single output bundle, in which case, declaration files were not correctly emitted
+  const output = await getCode(
+    bundle,
+    { format: 'esm', file: 'fixtures/noForceEmit/noEmit/dist/main.js' },
+    true
+  );
+
+  t.deepEqual(
+    output.map((out) => out.fileName),
+    // no `main.d.ts`, main.js is passed through
+    ['main.js']
+  );
+  t.is(warnings.length, 0);
+  // test that NO transpilation happened
+  const originalCode = fs.readFileSync(path.join(__dirname, input), 'utf8');
+  t.is(output[0].code, originalCode);
 });