+++ packages/commonjs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-commonjs ChangeLog
 
+## v22.0.2
+
+_2022-08-05_
+
+### Bugfixes
+
+- fix: Exclude multi-line template strings from indent (#1229)
+
+## v22.0.1
+
+_2022-06-24_
+
+### Bugfixes
+
+- fix: Only proxy detected commonjs entry points (#1180)
+
 ## v22.0.0
 
 _2022-04-24_