+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rollup/plugin-typescript ChangeLog
 
+## v8.4.0
+
+_2022-08-23_
+
+### Features
+
+- feat: allow override of forced `noEmit` and `emitDeclarationOnly` compiler options (#1242)
+
+## v8.3.4
+
+_2022-07-28_
+
+### Bugfixes
+
+- fix: fix declaration file generation for single file outputs (#1201)
+
+## v8.3.3
+
+_2022-06-10_
+
+### Bugfixes
+
+- fix: allow tslib peerDependency to be optional for npm and pnpm (#1191)
+
 ## v8.3.2
 
 _2022-04-13_